Enhance tenant management by adding restore functionality and updating query parameters for active and deleted states

+++ src/frontend/src/views/tenants/components/DataTable.vue
@@ -28,24 +28,29 @@ defineEmits(["update:options"]);
 
 const headers = ref([
   { title: t("admin.tenants.fields.isActive"), key: "isActive", align: "center", sortable: false, width: "128px" },
-  { title: t("admin.tenants.fields.name"), key: "name", sortable: false },
+  { title: t("admin.tenants.fields.name"), key: "name", sortable: true },
   { title: t("admin.tenants.fields.adminEmail"), key: "adminEmail", sortable: false },
-  { title: t("admin.tenants.fields.validUntil"), key: "validUntil.displayDateWithTime", align: "center", sortable: true },
+  { title: t("admin.tenants.fields.validUntil"), key: "validUntil", align: "center", sortable: true },
   { title: t("admin.tenants.fields.process"), key: "process", align: "center", sortable: false, width: "64px" }
 ]);
 
 const dataTableRef = ref(null);
 const operation = ref("delete");
+const recover = ref(false);
 const showModal = ref(false);
 const tenantId = ref(null);
 const confirmModalTitle = ref(null);
 const confirmModalMessage = ref(null);
 
 const handleConfirm = async () => {
-  if (operation.value === "delete") {
+  if (operation.value === "delete" && recover.value) {
     await tenantStore.delete(tenantId.value);
-  } else {
+  } else if (operation.value === "delete" && !recover.value) {
+    await tenantStore.restore(tenantId.value);
+  } else if (operation.value === "activate" && recover.value) {
     await tenantStore.activate(tenantId.value);
+  } else {
+    await tenantStore.deactivate(tenantId.value);
   }
 };
 
@@ -53,22 +58,23 @@ const handleCancel = () => {
   tenantId.value = null;
 };
 
-const showConfirmModal = (id, opt, value) => {
-  let activateTitle = null;
-  let activateMessage = null;
+const showConfirmModal = (id, opt, rec) => {
+  let modalTitle = null;
+  let modalMessage = null;
 
   if (opt === "activate") {
-    activateTitle = value ? t("admin.tenants.deactivate.title") : t("admin.tenants.activate.title");
-    activateMessage = value ? t("admin.tenants.deactivate.message") : t("admin.tenants.activate.message");
+    modalTitle = rec ? t("admin.tenants.deactivate.title") : t("admin.tenants.activate.title");
+    modalMessage = rec ? t("admin.tenants.deactivate.message") : t("admin.tenants.activate.message");
   } else {
-    activateTitle = t("admin.tenants.delete.title");
-    activateMessage = t("admin.tenants.delete.message");
+    modalTitle = rec ? t("admin.tenants.delete.title") : t("admin.tenants.restore.title");
+    modalMessage = rec ? t("admin.tenants.delete.message") : t("admin.tenants.restore.message");
   }
 
-  confirmModalTitle.value = activateTitle;
-  confirmModalMessage.value = activateMessage;
+  confirmModalTitle.value = modalTitle;
+  confirmModalMessage.value = modalMessage;
 
   operation.value = opt;
+  recover.value = rec;
   tenantId.value = id;
   showModal.value = true;
 };
@@ -91,17 +97,30 @@ const showConfirmModal = (id, opt, value) => {
         <v-icon size="small" :icon="item.isActive ? `$thumbUp` : `$thumbDown`" :color="item.isActive ? `success` : `error`" />
       </template>
 
+      <template #[`item.validUntil`]="{ item }">
+        {{ item.validUntil.displayDate }}
+      </template>
+
       <template #[`item.process`]="{ item }">
         <router-link :to="{ name: 'edit-tenant', params: { id: item.id } }">
           <v-icon size="small" icon="$pencil" color="secondary" />
         </router-link>
         <v-icon
-          v-show="item.isActive"
+          v-if="item.deleted"
+          size="small"
+          icon="$restore"
+          color="error"
+          class="ml-2"
+          @click="showConfirmModal(item.id, 'delete', false)"
+          :title="t('admin.tenants.restore.title')"
+        />
+        <v-icon
+          v-else
           size="small"
           icon="$trashCan"
           color="error"
           class="ml-2"
-          @click="showConfirmModal(item.id, `delete`, true)"
+          @click="showConfirmModal(item.id, 'delete', true)"
           :title="t('admin.tenants.delete.title')"
         />
         <v-icon
@@ -110,7 +129,7 @@ const showConfirmModal = (id, opt, value) => {
           size-="small"
           color="warning"
           class="ml-2"
-          @click="showConfirmModal(item.id, `activate`, true)"
+          @click="showConfirmModal(item.id, 'activate', true)"
           :title="t('admin.tenants.deactivate.title')"
         />
         <v-icon
@@ -120,7 +139,7 @@ const showConfirmModal = (id, opt, value) => {
           size-="small"
           color="success"
           class="ml-2"
-          @click="showConfirmModal(item.id, `activate`, false)"
+          @click="showConfirmModal(item.id, 'activate', false)"
           :title="t('admin.tenants.activate.title')"
         />
       </template>
